make the detailed examine verb predicted (#35362)
can we stop predicting things
