--- Content.Server/DetailExaminable/DetailExaminableComponent.cs
@@ -1,9 +1,0 @@
-﻿namespace Content.Server.DetailExaminable
-{
-    [RegisterComponent]
-    public sealed partial class DetailExaminableComponent : Component
-    {
-        [DataField("content", required: true)] [ViewVariables(VVAccess.ReadWrite)]
-        public string Content = "";
-    }
-}
